🐛 Fix: Split benchmarks module to separate client and server code
Build was failing because lib/benchmarks.ts imported 'fs' (Node.js module) but
was being imported by client components, causing Turbopack to fail.

Solution:
- Split into lib/benchmarks/types.ts (shared types)
- lib/benchmarks/utils.ts (client-safe utilities)
- lib/benchmarks/data.ts (server-only data loading with fs)
- lib/benchmarks/index.ts (convenience re-exports)

Components now explicitly import from types/utils, keeping fs isolated to
server-side code in app/benchmarks/page.tsx.

Fixes production build while maintaining clean type safety and DRY utilities.
