+++ lib/benchmarks/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Benchmark module - convenience re-exports
+ */
+
+export * from "./types";
+export * from "./utils";
+export * from "./data";
